+++ modules/versions.lua
@@ -6,7 +6,7 @@ return {
   ["Gen151"] = "1.5.2",
   ["Gen1AutoContinue"] = "1.2.2",
   ["Gen1AutoSave"] = "1.15.0",
-  ["Gen1Follower"] = "1.5.0",
+  ["Gen1Follower"] = "1.5.1",
   ["Gen1MenuManager"] = "0.3.2",
   ["Gen1ModMenu"] = "0.9.0",
   ["Gen1Remember"] = "1.0.1",